refactor(FIN-033): update import paths and clean up styles across multiple files
