+++ src/client/shared/styles/auth-image-container.scss
@@ -1,6 +1,5 @@
 .auth-image-container {
   display: none;
-
 }
 
 @media screen and (width >= 640px) {
@@ -15,5 +14,4 @@
     display: flex;
     min-width: 65%;
   }
-
 }
