+++ src/client/shared/styles/design-keys/light.scss
@@ -89,5 +89,3 @@
   --aqua-muted: rgb(223 247 249);
   --aqua-muted-foreground: rgb(21 94 117);
 }
-
-
