+++ src/client/shared/styles/globals.scss
@@ -20,6 +20,6 @@ html {
     display: none;
   }
 
-  -ms-overflow-style: none;  /* IE та Edge */
-  scrollbar-width: none;  /* Firefox */
+  -ms-overflow-style: none; /* IE та Edge */
+  scrollbar-width: none; /* Firefox */
 }
