+++ src/client/shared/styles/theme.scss
@@ -77,7 +77,6 @@
   --color-teal-foreground: var(--teal-foreground);
   --color-teal-muted: var(--teal-muted);
   --color-teal-muted-foreground: var(--teal-muted-foreground);
-
   --color-aqua: var(--aqua);
   --color-aqua-foreground: var(--aqua-foreground);
   --color-aqua-muted: var(--aqua-muted);
