+++ src/server/entities/country/infrastructure/country.orm.ts
@@ -1,5 +1,5 @@
 import { Column, Entity, OneToMany } from 'typeorm';
-import { DefaultTableColumnsOrm } from '../../../shared/infrastructure/default-table-columns.orm';
+import { DefaultTableColumnsOrm } from '../../../database/default-table-columns.orm';
 import { type CountryAndLocale } from '@common/records/countries.record';
 import { UserOrm } from '@backend/entities/user/infrastructure/user.orm';
 import { CountryRequirementsConstant } from '@common/constants/country-requirements.constant';
